create an authenticated lawyer
